Add files via upload
R is often criticised for not having simple and consistent pipelines for machine learning (ML) similar to those seen in Python. Here I am going to show a quick classification pipeline using the preprocessing capabilities of ‘recipes’ and ‘Biocomb’ for classification loops. This is a brief overview which can be used a series of different workflows. It can also be found on R Pubs at: https://rpubs.com/JakeTuricchi/classification_pipeline_R